move Migrator to separate library
