+++ libraries/migrations-manager/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    id("java")
+}
+
+group = "net.civmc.migratior"
+version = "1.0.0"
+
+dependencies {
+
+}
