+++ libraries/name-api/src/main/java/net/civmc/nameapi/NameAPI.java
@@ -10,6 +10,7 @@ import java.util.UUID;
 import javax.sql.DataSource;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
+import net.civmc.migratior.Migrator;
 
 public class NameAPI {
