+++ settings.gradle.kts
@@ -14,6 +14,7 @@ plugins {
 include(":ansible")
 
 include(":libraries:name-api")
+include(":libraries:migrations-manager")
 
 include(":plugins:announcements-velocity")
 include(":plugins:banstick-paper")
